# docs/swagger.yaml
basePath: /
definitions:
  commons.Metadata:
    properties:
      created_at:
        type: integer
      created_by:
        type: string
      deleted_at:
        type: integer
      deleted_by:
        type: string
      updated_at:
        type: integer
      updated_by:
        type: string
    type: object
  commons.Pagination:
    properties:
      page_number:
        type: integer
      page_size:
        type: integer
      total_item:
        type: integer
      total_pages:
        type: integer
    type: object
  commons.Response:
    properties:
      message:
        type: string
    type: object
...
  models.CreateUser:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.LoginUser:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  models.UpdateUser:
    properties:
      avatar_url:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
  userservice.JwtToken:
    properties:
      refresh_token:
        type: string
      token:
        type: string
    type: object
  userservice.OAuthInitiateResponse:
    properties:
      auth_url:
        type: string
    type: object
  userservice.User:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      metadata:
        $ref: '#/definitions/commons.Metadata'
      password:
        type: string
      roles:
        items:
          type: string
        type: array
    type: object
  userservice.UsersResponse:
    properties:
      pagination:
        $ref: '#/definitions/commons.Pagination'
      users:
        items:
          $ref: '#/definitions/userservice.User'
        type: array
    type: object
...
host: localhost:9000
info:
  contact:
    email: orkys.com@gmail.com
    name: API Support
    url: http://www.example.com/support
  description: This is the API documentation for the Orkys API Gateway
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: Orkys API
  version: "1.0"
paths:
...
  /list:
    get:
      consumes:
      - application/json
      description: Returns a list of users, limited by a query param (default 10).
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of events per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            $ref: '#/definitions/userservice.UsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List users
      tags:
      - user
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email and password and returns a JWT
        token.
      parameters:
      - description: Login credentials
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/models.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: Returns user information and tokens in cookies
          schema:
            $ref: '#/definitions/userservice.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized (invalid credentials)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Log in a user
      tags:
      - auth
  /oauth/{provider}/callback:
    get:
      consumes:
      - application/json
      description: Processes the OAuth callback from the provider, calls the user
        service, and returns a login response with JWT tokens.
      parameters:
      - description: OAuth provider (custom or google)
        in: path
        name: provider
        required: true
        type: string
      - description: State parameter for CSRF validation
        in: query
        name: state
        required: true
        type: string
      - description: Authorization code (and any other provider parameters)
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User login response with tokens in cookies
          schema:
            $ref: '#/definitions/userservice.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Process the OAuth callback from a provider
      tags:
      - auth
  /oauth/{provider}/initiate:
    get:
      consumes:
      - application/json
      description: Initiates an OAuth flow (for providers "custom" or "google") by
        calling the user service and returning the provider’s auth URL.
      parameters:
      - description: OAuth provider (custom or google)
        in: path
        name: provider
        required: true
        type: string
      - description: Optional state parameter. If omitted, a new state will be generated.
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OAuth initiation response
          schema:
            $ref: '#/definitions/userservice.OAuthInitiateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Initiate an OAuth flow for a provider
      tags:
      - auth
...
  /refresh:
    get:
      consumes:
      - application/json
      description: Takes a refresh token and returns a new access token (and possibly
        a new refresh token).
      parameters:
      - description: Bearer <refresh_token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: New access/refresh token pair in http-only cookies
          schema:
            $ref: '#/definitions/userservice.JwtToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized (missing or invalid token)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Refresh a JWT token
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new user in the system and returns the user object with
        tokens.
      parameters:
      - description: Registration details
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/models.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created user with tokens in cookies
          schema:
            $ref: '#/definitions/userservice.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict (user already exists)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /user:
    patch:
      consumes:
      - application/json
      description: Updates the user details (first name, last name, password and avatar
        URL) if they are not empty.
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: User details to update
        in: body
        name: UpdateUserRequest
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: Created user with tokens
          schema:
            $ref: '#/definitions/commons.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized (missing or invalid token)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update user details
      tags:
      - user
swagger: "2.0"